requirements.txt
langchain-community
smolagents
datasets
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
pypdf